feat: minor release v3.3.0; adds new backup capabilities (#14)
This PR adds:
- AWS Backup integration for all critical infrastructure (S3 buckets for CloudTrail, Grafana Loki, WAF, and ALB logs and EFS, EKS, and RDS).
-  Daily (at 2 AM UTC), weekly (at 3 AM UTC) and monthly (at 4 AM UTC) backup jobs that run independently of one another that are retained for 7 years.
- Adds backup.tf that defines the AWS Backup Vault for our deployment.
- Add a new dedicated KMS key for use with AWS Backup (total # of KMS keys 6 -> 7)
- Adds new terraform outputs for AWS Backup
- Improved destroy.sh to handle cases where there might be pending backup jobs or the vault might have recovery points.
- General improvements to destroy.sh
- Adds guards against failure modes where certain terraform show operations fail when the terraform state is partially or wholly deleted and/or corrupted
- Also adds more descriptive error handling and instructions for remediation for things like when the script fails because there’s no AWS credentials configured in the environment
- Changed the default number of verification attempts from 3 to 6 in restore.sh. This was done because in testing it was observed that while the verification almost always succeeded on the first attempt occasionally due to pod churn sometimes the verification takes two attempts. It’s never been observed in the many many runs it’s been through to ever need more than 2 attempts but just out of an abundance of caution I’ve decided to preemptively raise the number of attempts to 6 to guard against some statistically unlikely event that may or may not occur in the future while testing.
- Updated test-end-to-end-backup-restore.sh to use the destroy.sh script for the final destruction of infrastructure undertaken in the last step of the test.
- Previously this was accomplished through custom code which did pretty much the same thing as the destroy.sh script but it made little sense to keep maintaining this long term when the end to end test already uses the destroy.sh script earlier on for the initial destruction of infrastructure taken halfway through the test.
- Furthermore we should use the destroy script because it contains logic for handling the delete of the AWS Backup vault irregardless of what state the vault is in.
- Doing this guards against a potential failure mode where someone runs the end to end test too close to 2 AM UTC when the automated backup job goes off and then the terraform destroy would fail because the AWS Backup vault would have recovery points in it and or a pending backup job associated with it.
- Updated documentation to reflect the new functionality added in this PR.
- Also updates the version of the metrics server addon used by EKS

Overall this new backup capability provides the ability for users to do point in time recovery operations for specific components of the architecture at any time (i.e. restore only the EKS cluster to the state it was at this point in time, restore both the EKS cluster and the RDS to the state it was at this point in time, restore only the EFS to the state it was at this point in time, etc. etc.).  It also provides coverage, for the first time, to the four different s3 buckets we use for …
- CloudTrail events which audits AWS events in our accounts so we can analyze them to determine if anything undesired is occurring in our account,
- Application and node health logs stored in S3 via Grafana & Loki and scraped from our OpenEMR containers in real-time by Fluent Bit 4.1.1
- AWS Web Application Firewall events/logs (useful in the event there’s attempted cyberattacks or DDOS attacks or things to that affect; WAF would be our first line of defense in that case and should produce logs that indicated that something like that occurred and whether or not it was successful in mitigating the attack)
- AWS application load balancer logs showing network activity showing events associated with specific I.P. addresses for future analysis and auditing.

This is on top of the existing ability we had already with backup.sh and restore.sh that allows us to do on-demand backup and recovery both within or across regions in the current account or to other accounts entirely.

The original backup system focused on providing fast operational backups and restores for the cluster while taking into account the application logic of OpenEMR to ensure that the restore operation ends with your data restored and OpenEMR up and running. It's also in some ways more flexible than the AWS Backup based system as it allows for the exporting of a backup to another account. AWS Backup also allows for the functionality to export to another account but the requirement is that that target account be a part of the same Organization in AWS Organizations as the originating account. Granting the current backup system permissions to export to or restore to another account is in comparison a matter of IAM permissions in the originating and target account and does not require moving AWS Accounts between organizational units to function better.

For most realistic scenarios I would still recommend using the existing backup and restore system over the AWS Backup system where possible (note that this won’t be possible for anything involving S3 buckets which is outside of the scope of the original backup/restore system) because the restore.sh script was designed to take into account the specific application architecture of OpenEMR to ensure that after a restore operation OpenEMR is operational and you can open up the homepage and log in to the new setup. AWS Backup however is just going to rewind the state of whatever you tell it to to a specific moment in time. I’ll outline an example of when this would be an issue. Let’s say you set up the architecture and were using OpenEMR and uploading data to it for two days and then on the third day you changed your RDS password and sqlconf.php file and then shortly after that you had the terminal left open and you went back in accidentally deleted everything on the EFS. Alright, that's alright you have an automatic backup made the day before so you decide to rewind the EFS to the state it was in on day 2 with AWS Backup. After you were finished with this what you’d find is that OpenEMR would just return 500 errors and if you looked at application logs you’d see a lot of failed login attempts to MySQL because now you've restored to a version of sqlconf.php that no longer has valid credentials to access the database. While restore.sh knows that this could be a potential issue and takes this into account when conducting a restore AWS Backup is simply going to rewind components to a given point in time and it’s going to do so without regard for whether or not rewinding the component will actually result in a functionality installation at the end of the process.

Having said all of this it’s still good to have an additional set of backups that provides for point in time recovery, is not beholden to whether or not the cronjob setup to run backup.sh runs successfully and is backed by defined SLAs (https://docs.aws.amazon.com/aws-backup/latest/devguide/aws-backup-limits.html) from AWS.

This PR was tested against the test-end-to-end-backup-restore.sh script and passed.